+++ melt_the_code_dart/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.2.0
+
+- Implemented the useRuntimeQuery() use case function.
+
 ## 0.1.1
 
 - Some cleanup of documentation and exception changes to distinguish properly between the dart and flutter environments.